timerFunction/requirements.txt
azure-functions
requests
leanix_python_client-0.2.0-py3-none-any.whl
datetime